% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_dataset}
\alias{cc_dataset}
\title{Household Survey of 2004}
\format{Socio-economic and situational attributes used:
\describe{
  \item{Urb}{Urban Density}
  \item{Comp}{Household Composition }
  \item{Child}{Child category }
  \item{Day}{Day of the week, }
  \item{pAge}{Age category }
  \item{wstat}{Work status of an individual}
  \item{wstat}{Work status of partner}
  \item{SECx}{Socio-economic class}
  \item{Ncar}{Number of cars}
  \item{Gend}{Gender}
  \item{Driver}{Has a driving license}
  \item{Dur}{Duration}
}}
\usage{
cc_dataset
}
\description{
A dataset containing the interviews of 11000
 participants. The variables are as follows:
}
\keyword{datasets}

