% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchaid.R
\name{cc_define_file}
\alias{cc_define_file}
\title{This function reads user file}
\usage{
cc_define_file(x)
}
\arguments{
\item{x}{A file name}
}
\value{
A dataframe with user selected fields
}
\description{
This function reads user file
}
\examples{
cc_define_file("filename.txt")
}

