% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchaid.R
\name{cc_read_file}
\alias{cc_read_file}
\title{This function reads user file}
\usage{
cc_read_file(x)
}
\arguments{
\item{x}{A file name}
}
\value{
A dataframe with user selected fields in a tab-delimited text file.
}
\description{
This function reads user file
}
\examples{
my.df<-cc_read_file("filename.txt")
}

