% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_dataset}
\alias{cc_dataset}
\title{Household survey of 2004}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Urb}{Urban Residence}
  \item{Comp}{weight }
  \item{Child}{quality }
  \item{Day}{ colour, }
  \item{pAge}{a measurement  }

  \item{SECx}{length}
  \item{Ncar}{width }
  \item{Gend}{depth }
  \item{Driver}{total z }
  \item{Dur}{width of top }
}}
\usage{
cc_dataset
}
\description{
A dataset containing the interviews of 54,000
 participants The variables are as follows:
}
\keyword{datasets}

